refactor code from resolvers to correctly sit in controllers for file upload / delete actions
