+++ server/xpub-controller/index.js
@@ -1,9 +1,11 @@
 const SupportingFiles = require('./supportingFiles')
 const Manuscript = require('./manuscript')
 const S3Storage = require('./s3Storage')
+const ScienceBeamApi = require('./scienceBeamApi')
 
 module.exports = {
   SupportingFiles,
   Manuscript,
   S3Storage,
+  ScienceBeamApi,
 }
