+++ server/xpub-model/entities/file/index.test.js
@@ -67,21 +67,4 @@ describe('File', () => {
       expect(file.status).toBe('CANCELLED')
     })
   })
-
-  describe('save()', () => {
-    it('calls validate() on related manuscript', async () => {
-      const manuscript = await new Manuscript({
-        createdBy: userId,
-      }).save()
-      expect(manuscript.fileStatus).toEqual('READY')
-      const file = await new File({
-        manuscriptId: manuscript.id,
-        filename: 'thisfile.txt',
-        url: '/an/url',
-      }).save(userId)
-      await manuscript.refresh()
-      expect(file.status).toBe('CREATED')
-      expect(manuscript.fileStatus).toEqual('CHANGING')
-    })
-  })
 })
