+++ server/xpub-server/entities/manuscript/resolvers/updateManuscript.js
@@ -1,8 +1,10 @@
 const config = require('config')
+const { User } = require('@elifesciences/xpub-model')
 const { Manuscript } = require('@elifesciences/xpub-controller')
 
 async function updateManuscript(_, { data }, { user }) {
-  const ms = new Manuscript(config, user)
+  const userUuid = await User.getUuidForProfile(user)
+  const ms = new Manuscript(config, userUuid)
   return ms.update(data)
 }
