+++ server/xpub-server/entities/manuscript/resolvers/uploadManuscript.test.js
@@ -12,11 +12,10 @@ const { createTables } = require('@pubsweet/db-manager')
 const mailer = require('@pubsweet/component-send-email')
 const startS3rver = require('../../../test/mock-s3-server')
 const { User, Manuscript } = require('@elifesciences/xpub-model')
-const scienceBeamApi = require('./scienceBeamApi')
 const { Mutation } = require('.')
 const { userData, badUserData } = require('./index.test.data')
 const replaySetup = require('../../../../../test/helpers/replay-setup')
-const { S3Storage } = require('@elifesciences/xpub-controller')
+const { S3Storage, ScienceBeamApi } = require('@elifesciences/xpub-controller')
 
 describe('Manuscripts', () => {
   const profileId = userData.identities[0].identifier
@@ -53,11 +52,17 @@ describe('Manuscripts', () => {
       const blankManuscript = new Manuscript()
       blankManuscript.createdBy = userId
       const manuscript = await blankManuscript.save()
-
+      const file = {
+        filename: 'manuscript.pdf',
+        stream: fs.createReadStream(
+          `${__dirname}/../../../../../test/fixtures/dummy-manuscript-2.pdf`,
+        ),
+        mimetype: 'application/pdf',
+      }
       await expect(
         Mutation.uploadManuscript(
           {},
-          { id: manuscript.id },
+          { id: manuscript.id, file, fileSize: 73947 },
           { user: badProfileId },
         ),
       ).rejects.toThrow('Manuscript not found')
@@ -128,7 +133,7 @@ describe('Manuscripts', () => {
         mimetype: 'application/pdf',
       }
 
-      jest.spyOn(scienceBeamApi, 'extractSemantics').mockRejectedValueOnce({
+      jest.spyOn(ScienceBeamApi, 'extractSemantics').mockRejectedValueOnce({
         error: {
           code: 'ETIMEDOUT',
           connect: false,
